+++ esbuild.js
@@ -15,10 +15,24 @@ const extensionConfig = {
   external: ["vscode"],
 };
 
+const webviewConfig = {
+  minify: true,
+  entryPoints: ["./editor/webview.js"],
+  outfile: "./view/webview.js",
+};
+
+const webviewCssConfig = {
+  minify: true,
+  entryPoints: ["./editor/webview.css", "./editor/sortable-base.min.css"],
+  outdir: "./view",
+};
+
 (async () => {
   const args = process.argv.slice(2);
   try {
     await build(extensionConfig);
+    await build(webviewConfig);
+    await build(webviewCssConfig);
     console.log("build complete");
   } catch (err) {
     process.stderr.write(err);